добавила функции 2

+++ Python/functions_1/4.py
@@ -2,6 +2,6 @@ def min_num(num):
     return min(num)
 
 
-user_input=list(map(int, input('Введите пять чисел через пробел: ').split()))
+user_input=list(map(float, input('Введите пять чисел через пробел: ').split()))
 
 print('Минимальное число: ', min_num(user_input))
